+++ Tuple.cpp
@@ -1,7 +1,7 @@
 #include "Tuple.h"
 
 Tuple::Tuple(std::vector<Parameter *> input) {
-    for (int i = 0; i < input.size(); i++) {
+    for (long unsigned int i = 0; i < input.size(); i++) {
         values.push_back(input[i]->paramString());
     }
 }
